+++ CHANGELOG.md
@@ -3,12 +3,15 @@
 All notable changes to the "vscode-kusto-workbench" extension will be documented in this file.
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [0.4.0] - Unreleased
+## [0.5.0] - Unreleased
+
+
+## [0.4.0] - 2026-22-01
 
 * Added extension settings for controlling whether the extension opens .kql and .cls files
 * Copilot inline auto-complete (ghost text) support + new toolbar button turning it on / off
 * Bug fixes
-    * When a single editor contained multiple queries, it would complain about the non-active statements when they should be getting completely ignored. 
+    * When a single editor contained multiple queries, it would complain about the non-active statements when they should be getting completely ignored.
 
 ## [0.3.0] - 2026-15-01